% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpts_summary.R
\name{build_path_grobs}
\alias{build_path_grobs}
\title{Build grob list for arranging path plots.}
\usage{
build_path_grobs(cpts_path)
}
\arguments{
\item{cpts_path}{Head of path.}
}
\value{
List of grobs.
}
\description{
Build grob list for arranging path plots.
}
