% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpts_summary.R
\name{build_path_layout}
\alias{build_path_layout}
\title{Build layout matrix for arranging path grobs.}
\usage{
build_path_layout(cpts_path, plot_num)
}
\arguments{
\item{cpts_path}{Head of path.}

\item{plot_num}{Current plot number (start with 1).}
}
\value{
Layout matrix for arrangeGrob.
}
\description{
Build layout matrix for arranging path grobs.
}
