% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{create_spec_count_rows}
\alias{create_spec_count_rows}
\title{Function to create rows for missing 0's in yr_dat}
\usage{
create_spec_count_rows(
  Island = NA,
  Year = NA,
  Observer = "UNK",
  Station = NA,
  X = NA,
  Y = NA,
  Date = NA,
  Start = NA,
  Species_Code = NA,
  Count = NA
)
}
\arguments{
\item{Island}{String - island of observation.}

\item{Year}{Integer - year of observation}

\item{Observer}{String - initials of observer or "UNK" for unknown}

\item{Station}{String - station of observation}

\item{X}{Numeric - x coord of station}

\item{Y}{Numeric - y coord of station}

\item{Date}{Date - date of observation}

\item{Start}{Time - start time of survey}

\item{Species_Code}{Character vector with all species codes of birds to create rows for}

\item{Count}{Integer - number of birds counted, typically 0 in this usage}
}
\value{
Dataframe with new rows for adding into BBS dataframe.
}
\description{
Function to create rows for missing 0's in yr_dat
}
