% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_model.R
\name{dep_check}
\alias{dep_check}
\title{Check dependencies when building formula strings.}
\usage{
dep_check(combo, dep)
}
\arguments{
\item{combo}{Character vector with combination of terms}

\item{dep}{Dataframe with dependencies to check}
}
\value{
Character vector with appropriate term dependencies
}
\description{
Check dependencies when building formula strings.
}
