% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_summary.R
\name{error_comp_rf}
\alias{error_comp_rf}
\title{Calculate MAPE and RMSE for rf and benchmark glm.}
\usage{
error_comp_rf(rf_out)
}
\arguments{
\item{rf_out}{Dataframe from model_rf.}
}
\value{
Tibblewith train/test MAPE and RMSE for each model.
}
\description{
Calculate MAPE and RMSE for rf and benchmark glm.
}
