% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_summary.R
\name{error_table_rf}
\alias{error_table_rf}
\title{Calculate MAPE and RMSE for rf and benchmark glm and format into gt table.}
\usage{
error_table_rf(rf_out)
}
\arguments{
\item{rf_out}{Dataframe from model_rf.}
}
\value{
gt table with train/test MAPE and RMSE for each model.
}
\description{
Calculate MAPE and RMSE for rf and benchmark glm and format into gt table.
}
