% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_counts.R
\name{filter_birds}
\alias{filter_birds}
\title{Convenient wrapper for filtering bird species in BBS data.}
\usage{
filter_birds(bbs_dat, active_birds)
}
\arguments{
\item{bbs_dat}{Dataframe with BBS data.}

\item{active_birds}{Character vector of species codes to retain.}
}
\value{
Dataframe filtered to only include active birds.
}
\description{
Convenient wrapper for filtering bird species in BBS data.
}
