% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_counts.R
\name{filter_stations}
\alias{filter_stations}
\title{Convenient wrapper for filtering stations in BBS data.}
\usage{
filter_stations(bbs_dat, active_stations)
}
\arguments{
\item{bbs_dat}{Dataframe with BBS data.}

\item{active_stations}{Character or numeric vector with stations to retain.}
}
\value{
Dataframe filtered to only include active stations.
}
\description{
Convenient wrapper for filtering stations in BBS data.
}
