% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpts_model.R
\name{fit_cpt}
\alias{fit_cpt}
\title{Given data and location of changepoint, fit a model.}
\usage{
fit_cpt(cpt, bdat)
}
\arguments{
\item{cpt}{Changepoint to fit. Should be a value in time column.}

\item{bdat}{Dataframe with time and count.}
}
\value{
glm object for changepoint model.
}
\description{
Given data and location of changepoint, fit a model.
}
