% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_model.R
\name{fit_glm}
\alias{fit_glm}
\title{Fit a Poisson or Negative Binomial GLM to determine}
\usage{
fit_glm(bdat, formulas, family)
}
\arguments{
\item{bdat}{BBS data filtered to one specific species. Dataframe with columns YrQtr, Year, Year2, Qtr, N_Stations, and Count.}

\item{formulas}{Character vector with model formulas.}

\item{family}{Type of model to use. One of "poisson" or "negative binomial".}
}
\value{
Object of class glm, with formula selected via AIC.
}
\description{
Fit a Poisson or Negative Binomial GLM to determine
}
