% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{fix_date}
\alias{fix_date}
\title{Format dates that aren't already of the standard forms "\%Y-\%m-\%d" or "\%Y/\%m/\%d"}
\usage{
fix_date(dat)
}
\arguments{
\item{dat}{Dataframe with BBS year data to be corrected.}
}
\value{
Same dataframe as input, but correctly formatted date column
}
\description{
Format dates that aren't already of the standard forms "\%Y-\%m-\%d" or "\%Y/\%m/\%d"
}
