% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{fix_date_2021}
\alias{fix_date_2021}
\title{Fix inconsistent date formatting for 2021}
\usage{
fix_date_2021(dat)
}
\arguments{
\item{dat}{Dataframe with 2021 BBS year data}
}
\value{
Same dataframe as input, but correctly formatted date column
}
\description{
Fix inconsistent date formatting for 2021
}
