% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpts_model.R
\name{get_cpts}
\alias{get_cpts}
\title{Identify possible changepoints for testing.}
\usage{
get_cpts(bdat, min_len = 10)
}
\arguments{
\item{bdat}{Dataframe with time and count.}

\item{min_len}{Minimum time length for any segment.}
}
\value{
Vector of valid changepoints to test.
}
\description{
Identify possible changepoints for testing.
}
