% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_model.R
\name{get_glm_formula}
\alias{get_glm_formula}
\title{Extract formula character vector from glm object}
\usage{
get_glm_formula(mod)
}
\arguments{
\item{mod}{glm object}
}
\value{
character vector representing formula
}
\description{
Extract formula character vector from glm object
}
