% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_island_year}
\alias{get_island_year}
\title{Read data for a given island and year}
\usage{
get_island_year(db_file, db_names, year_ranges, isl, yr)
}
\arguments{
\item{db_file}{Path to the BBS .xlsx database file.}

\item{db_names}{Column names in BBS database file.}

\item{year_ranges}{Dataframe of the kind returned by get_year_rows.}

\item{isl}{One of "Saipan", "Rota", "Tinian".}

\item{yr}{Year to retrieve as an integer.}
}
\value{
Dataframe containing BBS observations from that island and year.
}
\description{
Read data for a given island and year
}
