% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_island_year_list}
\alias{get_island_year_list}
\title{Get all years for an island and store as a list - avoid slow re-reading from Excel}
\usage{
get_island_year_list(db_file, year_ranges, isl)
}
\arguments{
\item{db_file}{Path to the BBS .xlsx database file.}

\item{year_ranges}{Dataframe of the kind returned by get_year_rows.}

\item{isl}{One of "Saipan", "Rota", "Tinian".}
}
\value{
List where each entry is a dataframe with yearly data for that island.
}
\description{
Get all years for an island and store as a list - avoid slow re-reading from Excel
}
