% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpts_summary.R
\name{get_modeled_cpts}
\alias{get_modeled_cpts}
\title{Retrieve fitted changepoints from model list}
\usage{
get_modeled_cpts(modlist)
}
\arguments{
\item{modlist}{Model list from model_cpts}
}
\value{
Vector of changepoints.
}
\description{
Retrieve fitted changepoints from model list
}
