% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{get_n_stats}
\alias{get_n_stats}
\title{Get number of stations per survey from BBS data.}
\usage{
get_n_stats(bbs_dat)
}
\arguments{
\item{bbs_dat}{Dataframe with BBS data. Must have YrQtr and Station.}
}
\value{
Dataframe number of stations by YrQtr.
}
\description{
Get number of stations per survey from BBS data.
}
