% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpts_summary.R
\name{get_node_labels}
\alias{get_node_labels}
\title{Retrieve node (changepoint) labels from a changepoint fit path.}
\usage{
get_node_labels(path)
}
\arguments{
\item{path}{Path list from model_cpts or model_birds_cpts.}
}
\value{
List of labels for each node in the fit path.
}
\description{
Retrieve node (changepoint) labels from a changepoint fit path.
}
