% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpts_summary.R
\name{get_node_plt}
\alias{get_node_plt}
\title{Find a node_plt that corresponds to a node_label}
\usage{
get_node_plt(path, label)
}
\arguments{
\item{path}{Path list from model_cpts or model_birds_cpts.}

\item{label}{Label corresponding node plot to retrieve. Should come from get_node_labels.}
}
\value{
node_plt grob or NULL if no such label.
}
\description{
Find a node_plt that corresponds to a node_label
}
