% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observers.R
\name{get_observers}
\alias{get_observers}
\title{Gather observers over time}
\usage{
get_observers(bbs_dat)
}
\arguments{
\item{bbs_dat}{Dataframe with BBS data. Must have columns YrQtr, Observer, and Replicate}
}
\value{
Dataframe with year-quarter, observer's initials, chronological order, and number of surveys conducted.
}
\description{
Gather observers over time
}
