% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_rows}
\alias{get_rows}
\title{Read rows from database. Used to retrieve contiguous survey segments}
\usage{
get_rows(db_file, db_names, year_range)
}
\arguments{
\item{db_file}{Path to the BBS .xlsx database file.}

\item{db_names}{Column names in BBS database file.}

\item{year_range}{Dataframe with 1 row of from get_year_rows. Specifies island, year, and corresponding rows in database file.}
}
\value{
Dataframe containing BBS observations from that island and year.
}
\description{
Read rows from database. Used to retrieve contiguous survey segments
}
