% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_times.R
\name{get_survey_end_times}
\alias{get_survey_end_times}
\title{Gather latest survey start times in relation to sunrise.}
\usage{
get_survey_end_times(bbs_dat)
}
\arguments{
\item{bbs_dat}{Dataframe with BBS data. Must have columns Year, Station, Date,}
}
\value{
Dataframe with YrQtr, start time in minutes from sunrise, and bins for more/less than 30 minutes before/after 3 hours from sunrise.
}
\description{
Gather latest survey start times in relation to sunrise.
}
