% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_year_rows}
\alias{get_year_rows}
\title{Get rows for each year}
\usage{
get_year_rows(db_file, sheet)
}
\arguments{
\item{db_file}{Path to the BBS .xlsx database file.}

\item{sheet}{Sheet to read. Either string with island name or integer.}
}
\value{
Dataframe with island, year, and first/last rows in Excel sheet corresponding to that year.
}
\description{
Get rows for each year
}
