% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_summary.R
\name{glm_exec_summary}
\alias{glm_exec_summary}
\title{Create a summary table indicating trends identified in the models.}
\usage{
glm_exec_summary(bmod_tbl)
}
\arguments{
\item{bmod_tbl}{Tibble output of model_birds_glm.}
}
\value{
gt table summarizing trends, model type, and model fit.
}
\description{
Create a summary table indicating trends identified in the models.
}
