% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_summary.R
\name{glm_exec_summary_row}
\alias{glm_exec_summary_row}
\title{Create row of executive summary table.}
\usage{
glm_exec_summary_row(bmod)
}
\arguments{
\item{bmod}{Object of class glm or negbin corresponding to model for one species.}
}
\value{
Dataframe with one row and columns Year, Season, and N_Stations, Family, and GOF for building into executive summary table.
}
\description{
Create row of executive summary table.
}
