% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_model.R
\name{gof_test}
\alias{gof_test}
\title{Goodness of fit test based on residual deviance.}
\usage{
gof_test(glm)
}
\arguments{
\item{glm}{glm object to conduct test for.}
}
\value{
List of 3 consisting of residual deviance, degrees of freedom, and p-value.
}
\description{
Goodness of fit test based on residual deviance.
}
