% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_counts.R
\name{islandwide_counts}
\alias{islandwide_counts}
\title{Sum counts across stations for island-wide totals}
\usage{
islandwide_counts(bbs_dat)
}
\arguments{
\item{bbs_dat}{Dataframe with BBS data.}
}
\value{
Dataframe with YrQtr, Species_Code, and Count.
}
\description{
Sum counts across stations for island-wide totals
}
