% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpts_model.R
\name{model_birds_cpts}
\alias{model_birds_cpts}
\title{Build changepoint models and plots for multiple species.
Convenient wrapper for applying to multiple species at once. Modeling driven by model_cpts and plotting done by plot_cpts_model.}
\usage{
model_birds_cpts(
  dat,
  min_len = 10,
  alpha = 0.05,
  multiplicity = "BH",
  boot_iter = 10^3,
  robust_se = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dat}{Dataframe with time and count.}

\item{min_len}{Minimum time length (data points) to model changepoints.}

\item{alpha}{Specify level at which to control FWE or FDR.}

\item{multiplicity}{Specify form of multiplicity control. One of "none", "bonferroni", "BH" (equivalently, "FDR"), "BY". Defaults to "BH"}

\item{boot_iter}{Number of bootstrap iterations for changepoint selection statistic.}

\item{robust_se}{Flag for including HAC robust standard errors from package 'sandwich' in plots.}

\item{verbose}{Flag for printing timing and progress updates.}
}
\value{
A tibble with species_code (char), models (list of list of negbin objects), paths (list of path lists), fit_plots, resid_plots, and unexp_plots (all lists of ggplots)
}
\description{
Build changepoint models and plots for multiple species.
Convenient wrapper for applying to multiple species at once. Modeling driven by model_cpts and plotting done by plot_cpts_model.
}
