% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_model.R
\name{model_birds_glm}
\alias{model_birds_glm}
\title{Fit GLMs for one or more species based on island-wide counts.}
\usage{
model_birds_glm(
  dat,
  family = "poisson",
  species_formulas = NULL,
  response = "Count",
  choose_terms = c("Year", "Year2", "Qtr"),
  fix_terms = "N_Stations",
  depends = tibble(term = "Year2", on = "Year"),
  robust_se = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dat}{Dataframe with counts and predictors for each species to be modeled}

\item{family}{Family of GLM to apply to data. One of "poisson", "negative binomial", or "auto".}

\item{species_formulas}{Optional - tibble with two columns: Species_Code and formulas. Each entry of formulas should be a list of valid formula character vectors.}

\item{response}{Optional - name of response column. Default is Count.}

\item{choose_terms}{Optional - names of terms selected by AIC. Defaults are Year, Year2, and Qtr.}

\item{fix_terms}{Optional - names of terms always included in model. Defaults to N_Stations.}

\item{depends}{Optional - tibble with two columns: term and on. term is a name in choose_terms and is another name from choose_terms that term depends on. Default has term = Year2, on = Year.}

\item{robust_se}{Flag for including HAC robust standard errors from package 'sandwich' in plots.}

\item{verbose}{Flag for printing timing and progress updates.}
}
\value{
A tibble with species_code (char), models (list of glm objects), fit_plots, back_trans_plots, resid_plots, and unexp_plots (all lists of ggplots)
}
\description{
Fit GLMs for one or more species based on island-wide counts.
}
