% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpts_model.R
\name{model_cpts}
\alias{model_cpts}
\title{Fit a glm model model with an arbitrary number of changepoints.
Applies binary segmentation to obtain changepoints that approximately maximize likelihood.}
\usage{
model_cpts(
  bdat,
  min_len = 10,
  alpha = 0.05,
  multiplicity = "BH",
  boot_iter = 10^3
)
}
\arguments{
\item{bdat}{Dataframe with time and count.}

\item{min_len}{Minimum time length to model changepoints.}

\item{alpha}{Specify level at which to control FWE or FDR.}

\item{multiplicity}{Specify form of multiplicity control. One of "none", "bonferroni", "BH" (equivalently, "FDR"), "BY". Defaults to "BH"}

\item{boot_iter}{Number of bootstrap iterations for changepoint selection statistic.}
}
\value{
List of glm objects collectively representing multiple-changepoint model.
}
\description{
Fit a glm model model with an arbitrary number of changepoints.
Applies binary segmentation to obtain changepoints that approximately maximize likelihood.
}
