% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_model.R
\name{model_rf}
\alias{model_rf}
\title{Build a forecasting model for counts using random forest, benchmarked against our standard GLM approach.
Forecasting for random forest is performed recursively so that we can compare models over the training data.}
\usage{
model_rf(
  dat,
  train_start,
  test_start,
  forecast_start,
  horizon,
  lag_order,
  seed = NULL
)
}
\arguments{
\item{dat}{Dataframe for a single species as prepared by prep_dat_glm.}

\item{train_start}{First YrQtr for train set.}

\item{test_start}{First YrQtr for test set.}

\item{forecast_start}{First YrQtr for forecast set.}

\item{horizon}{Number of survey periods to forecast.}

\item{lag_order}{Number of lags to include in model.}

\item{seed}{Optional seed to set so that random forest results can be replicated.}
}
\value{
Dataframe with random forest train/test predictions and glm benchmark train/test predictions.
}
\description{
Build a forecasting model for counts using random forest, benchmarked against our standard GLM approach.
Forecasting for random forest is performed recursively so that we can compare models over the training data.
}
