% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpts_summary.R
\name{pad_layout}
\alias{pad_layout}
\title{Pad layout matrix with NAs (for balancing layout).}
\usage{
pad_layout(lay, height)
}
\arguments{
\item{lay}{Layout matrix to pad.}

\item{height}{Height to pad to.}
}
\value{
Padded layout matrix.
}
\description{
Pad layout matrix with NAs (for balancing layout).
}
