% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_counts.R
\name{plot_counts}
\alias{plot_counts}
\title{Create plots for multiple bird species. If no species specified, plot for all available.}
\usage{
plot_counts(bbs_dat, birds = c())
}
\arguments{
\item{bbs_dat}{Dataframe with BBS data.}

\item{birds}{Character vector with bird species codes.}
}
\value{
List of ggplots for each bird species in birds.
}
\description{
Create plots for multiple bird species. If no species specified, plot for all available.
}
