% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpts_summary.R
\name{plot_cpts_dist}
\alias{plot_cpts_dist}
\title{Plot changepoint distribution over multiple species.}
\usage{
plot_cpts_dist(cpts_fits)
}
\arguments{
\item{cpts_fits}{Fitted models from model_birds_cpts.}
}
\value{
ggplot of stacked bar chart displaying the distribbution of changepoints.
}
\description{
Plot changepoint distribution over multiple species.
}
