% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpts_summary.R
\name{plot_cpts_model}
\alias{plot_cpts_model}
\title{Plot changepoint models.}
\usage{
plot_cpts_model(modlist, robust_se = FALSE, species = "")
}
\arguments{
\item{modlist}{Model list from model_cpts}

\item{robust_se}{Flag for including HAC robust standard errors from package 'sandwich' in plots.}

\item{species}{Species code for species being modeled. Used in title of plot.}
}
\value{
ggplot of fitted model.
}
\description{
Plot changepoint models.
}
