% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpts_summary.R
\name{plot_cpts_path}
\alias{plot_cpts_path}
\title{Plot binary segmentation path for changepoint model.}
\usage{
plot_cpts_path(cpts_path)
}
\arguments{
\item{cpts_path}{Path list from model_cpts or model_birds_cpts.}
}
\value{
Graphical object representing changepoint path.
}
\description{
Plot binary segmentation path for changepoint model.
}
