% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_times.R
\name{plot_end_times}
\alias{plot_end_times}
\title{Plot latest start times over time.}
\usage{
plot_end_times(end_dat)
}
\arguments{
\item{end_dat}{Start time dataframe from get_survey_end_times}
}
\value{
ggplot of start times over time.
}
\description{
Plot latest start times over time.
}
