% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{plot_n_stats}
\alias{plot_n_stats}
\title{Plot number of stations per survey over time}
\usage{
plot_n_stats(n_stats)
}
\arguments{
\item{n_stats}{Dataframe from get_n_stats.}
}
\value{
ggplot with number of stations surveyed over time.
}
\description{
Plot number of stations per survey over time
}
