% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observers.R
\name{plot_observers}
\alias{plot_observers}
\title{Generate plot of observers over time}
\usage{
plot_observers(obs_dat, filter_obs = FALSE)
}
\arguments{
\item{obs_dat}{Observer dataframe from get_obervers.}

\item{filter_obs}{Filter observers to only show those who have conducted more than one survey. Default does not filter.}
}
\value{
Plot of observers over time.
}
\description{
Generate plot of observers over time
}
