% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_summary.R
\name{plot_rf_model}
\alias{plot_rf_model}
\title{Build forecast plot.}
\usage{
plot_rf_model(rf_out)
}
\arguments{
\item{rf_out}{Dataframe from model_rf.}
}
\value{
ggplot with forecast models.
}
\description{
Build forecast plot.
}
