% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_times.R
\name{plot_start_times}
\alias{plot_start_times}
\title{Plot start times over time.}
\usage{
plot_start_times(start_dat)
}
\arguments{
\item{start_dat}{Start time dataframe from get_survey_start_times}
}
\value{
Plot of start times over time.
}
\description{
Plot start times over time.
}
