% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{plot_station_count}
\alias{plot_station_count}
\title{Create plot for individual station.}
\usage{
plot_station_count(station, bbs_dat)
}
\arguments{
\item{station}{Character vector for Station found in bdat.}

\item{bbs_dat}{BBS dataframe with data for 'station' Station. May have more as well, intended to be passed through lapply in plot_counts.}
}
\value{
ggplot of counts over time for bird.
}
\description{
Create plot for individual station.
}
