% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{plot_station_counts}
\alias{plot_station_counts}
\title{Create plots for multiple stations. If no station specified, plot for all available.}
\usage{
plot_station_counts(bbs_dat, stations = c())
}
\arguments{
\item{bbs_dat}{Dataframe with BBS data.}

\item{stations}{Character vector with stations.}
}
\value{
List of ggplots for each bird species in birds.
}
\description{
Create plots for multiple stations. If no station specified, plot for all available.
}
