% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_model.R
\name{prep_dat_glm}
\alias{prep_dat_glm}
\title{Prep BBS data for GLM modeling.
Create dataframe with Count, Year, Year squared, Quarter, and N_Stations for passing to model_birds_glm.}
\usage{
prep_dat_glm(bbs_dat)
}
\arguments{
\item{bbs_dat}{Dataframe with BBS data.}
}
\value{
Dataframe with columns suitable for BrBS' default glm modeling.
}
\description{
Prep BBS data for GLM modeling.
Create dataframe with Count, Year, Year squared, Quarter, and N_Stations for passing to model_birds_glm.
}
