% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_counts.R
\name{species_counts_summary}
\alias{species_counts_summary}
\title{Create a summary table for count data availability.}
\usage{
species_counts_summary(counts, order_col = "firstYr", descending = FALSE)
}
\arguments{
\item{counts}{Count dataframe from islandwide_counts.}

\item{order_col}{Character vector for column to sort table by. One of Species_Code, firstYr, lastYr, propSurveys, lastCount.}

\item{descending}{Logical to sort in descending order. Default is ascending.}
}
\value{
gt table with data availability info.
}
\description{
Create a summary table for count data availability.
}
