% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{station_counts}
\alias{station_counts}
\title{Sum counts across species for station-level total counts.}
\usage{
station_counts(bbs_dat)
}
\arguments{
\item{bbs_dat}{Dataframe iwht BBS data}
}
\value{
Dataframe with YrQtr, Station, and Count.
}
\description{
Sum counts across species for station-level total counts.
}
