% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpts_summary.R
\name{trim_layout}
\alias{trim_layout}
\title{Trim excess NAs from layout matrix.}
\usage{
trim_layout(lay)
}
\arguments{
\item{lay}{Layout matrix to trim.}
}
\value{
Trimmed layout matrix.
}
\description{
Trim excess NAs from layout matrix.
}
