% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_summary.R
\name{visualize_bird_glm}
\alias{visualize_bird_glm}
\title{Produce visualizations and model summary for a single species.}
\usage{
visualize_bird_glm(glm_out, bdx)
}
\arguments{
\item{glm_out}{Tibble output of model_birds_glm.}

\item{bdx}{Species code or integer index for lists in glm_out.}
}
\value{
Prints plots associated with species model.
}
\description{
Produce visualizations and model summary for a single species.
}
