% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{year_list_to_df}
\alias{year_list_to_df}
\title{Clean and combine list from get_island_year_list into one dataframe.}
\usage{
year_list_to_df(all_yrs)
}
\arguments{
\item{all_yrs}{List from get_island_year_list}
}
\value{
Dataframe from combining entries of all_yrs list, with additional column YrQtr
}
\description{
Clean and combine list from get_island_year_list into one dataframe.
}
