% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observers.R
\name{get_observers}
\alias{get_observers}
\title{Title}
\usage{
get_observers(all_yrs, yrs)
}
\arguments{
\item{all_yrs}{List of dataframes produced by get_island_year_list.}

\item{yrs}{Numeric vector of years to pull from all_yrs}
}
\value{
Dataframe with year-quarter, observer's initials, chronological order, and number of surveys conducted.
}
\description{
Title
}
