% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_times.R
\name{get_survey_start_times}
\alias{get_survey_start_times}
\title{Gather survey start times to use in relation to sunrise as a covariate}
\usage{
get_survey_start_times(all_yrs, yrs, isl)
}
\arguments{
\item{all_yrs}{List of dataframes produced by get_island_year_list.}

\item{yrs}{Numeric vector of years to pull from all_yrs}

\item{isl}{One of "Saipan", "Rota", "Tinian".}
}
\value{
Dataframe with year-quarter, start time in minutes from sunrise, and bins for more/less than 30 minutes before/after sunrise
}
\description{
Gather survey start times to use in relation to sunrise as a covariate
}
